Surface GLM (WIP)

+++ test/utest_import_nirs_in_bst.m
@@ -1,17 +1,28 @@
-function sFile = utest_import_nirs_in_bst(nirs_fn)
-%% Ensure that nst_utest protocol exists
-ProtocolName = 'nst_utest';
-% Delete existing protocol
-gui_brainstorm('DeleteProtocol', ProtocolName);
+function sFile = utest_import_nirs_in_bst(nirs_fn, clean)
 
-db_dir = bst_get('BrainstormDbDir');
-nst_protocol_dir = fullfile(db_dir, ProtocolName);
-if exist(nst_protocol_dir, 'dir')
-    rmdir(nst_protocol_dir, 's');
+if nargin < 2
+    clean = 1; % always clean by default
 end
 
-% Create new protocol
-gui_brainstorm('CreateProtocol', ProtocolName, 0, 0);
+ProtocolName = 'nst_utest';
+
+%% Clean nst_utest protocol if needed
+if clean && ~isempty(bst_get('Protocol', ProtocolName))
+    % Delete existing protocol
+    gui_brainstorm('DeleteProtocol', ProtocolName);
+    
+    db_dir = bst_get('BrainstormDbDir');
+    nst_protocol_dir = fullfile(db_dir, ProtocolName);
+    if exist(nst_protocol_dir, 'dir')
+        rmdir(nst_protocol_dir, 's');
+    end
+end
+
+%% Ensure that nst_utest protocol exists
+if isempty(bst_get('Protocol', ProtocolName))
+    % Create new protocol
+    gui_brainstorm('CreateProtocol', ProtocolName, 0, 0);
+end
 
 %% Import data as raw file
 [rr, bfn, ext] = fileparts(nirs_fn);
